Fix GitInfo generation when git executable not installed
